refactor: extract message components
